HuyDev: prototype home, Disable onTouch viewpager root, move classes to respective packages
